Records list directive srv. Refactoring
