+++ angular-spa-ui/js/admin-app/content-manager-module/index.js
@@ -5,15 +5,14 @@ module.exports = function (angular) {
     var contentManager = angular.module('app.contentManager', []);
 
 
-    var cmCtrl = require('./contentManager-ctrl'),
-        contentStorageSrv = require('./contentStorage-srv'),
-        contentManagerSrv = require('./contentManager-srv');
-    var testCtrl = require('./TestCtrl');
+    var cmCtrl = require('./contentManager-ctrl');
+    var contentStorageSrv = require('./contentStorage-srv');
+    var contentManagerSrv = require('./contentManager-srv');
+
 
     contentManager
         .constant('contentManagerConfig', require('./contentManager-config'))
         .controller('cmCtrl', cmCtrl)
-        .controller('TestCtrl', testCtrl)
         .service('contentStorage', contentStorageSrv)
         .service('cmService', contentManagerSrv)
         .config(configCb);
@@ -40,16 +39,6 @@ module.exports = function (angular) {
                     "module-content": {
                         templateUrl: '/admin.html',
                         controller: 'cmCtrl'
-                    }
-                }
-            })
-            .state('test', {
-                url: '/test',
-                views: {
-                    "module-content": {
-                        template: 'This is test url! {{test.content}} {{test.model.qwerty}}<div ng-click="test.privateApi.tttt(\'Our console log or alert text\')">Click here for this function</div><div ng-click="aaaa(\'Our console log or alert text\')">Click here for scope function</div>',
-                        controller: 'TestCtrl',
-                        controllerAs: 'test'
                     }
                 }
             });
